Specify the correct calling convention (stdcall) for NewWerReportSubmit()
This function was previously declared with the default calling convention (cdecl), which was causing
WerFault.exe to sometimes crash with an access violation error (exit code 0xc0000005) when it was
launched for a crashed 32-bit process. This was due to the stack getting corrupted. This didn't
occur for crashes in 64-bit processes because these invoke the 64-bit WerFault.exe and both stdcall
and cdecl resolve to the same calling convention (Microsoft x64) on 64-bit.