Let WerTweakInject check the address of the debug breakpoint in TranslateSnapshotHandleByDebugger()
This is being done by putting TranslateSnapshotHandleByDebugger() in a separate, dedicated PE
section of WerTweak32.dll and WerTweak64.dll. When WerTweakInject starts up, it first loads both
DLLs into memory and saves the RVA and size of this PE section for each DLL. Additionally, when the
injected LoadLibrary() stub has finished executing, WerTweakInject saves the base address of the DLL
injected into WerFault.exe. By combining this base address with the RVA and size of the PE section,
WerTweak can now check whether a breakpoint occurs in this section, and hence in
TranslateSnapshotHandleByDebugger().

Also added a new "Retrieve info of DLLs to inject" command to WerTweakTest.