Translate process snapshot handles passed to PssQuerySnapshot() and WerpTraceSnapshotStatistics()
PssQuerySnapshot() was already hooked when called from WerFault.exe but not when called from wer.dll, the latter is also required on more recent Windows 11 Insider Preview builds, otherwise WerFault.exe fails again with exit code 0xc0000008 (invalid handle) when it terminates a hung process.
WerpTraceSnapshotStatistics() is an undocumented wer.dll function that is called from WerFault.exe, also only on (more recent) preview builds. One of its parameters is the process snapshot handle that was created by WerSvc. This function explicitly checks if the memory block corresponding to the snapshot handle starts with the process snapshot signature (PSSNT_SIGNATURE_PSSD), so this was causing WerFault.exe to fail with 0xc0000005 (access violation) when it tried to read the memory of the original snapshot handle rather than the translated handle.