Translate process snapshot handles passed to PssWalkSnapshot()
This fixes another instance of WerFault.exe failing with exit code 0xc0000008 (invalid handle) when it terminates a hung process on more recent Windows 11 Insider Preview builds.

Note: these PssWalkSnapshot() calls only seem to occur when WerFault is running while there is no connection to the Internet (probably related to the fact that this means that the problem report has to be queued locally first). Additionally, the AllowTelemetry registry setting must be set to value 3 (full level telemetry).

Note 2: This fix is currently not practical due to the sheer amount of times that WerFault calls PssWalkSnapshot() (several thousands). WerTweak's handle translation together with its DbgOut() calls make WerFault way too slow to be able to finish its work in a reasonable amount of time. This will be fixed in subsequent commit(s).