Trigger process snapshot handle translation with a custom exception instead of a breakpoint
This custom exception uses an application-defined exception code STATUS_TRANSLATE_PROCESS_SNAPSHOT_HANDLE (0x62DA1ECE). Even though this is now less important than with a breakpoint, we still verify that the exception address is in TranslateSnapshotHandleByDebugger().

This now allows us to pass the address of the snapshot handle as an exception parameter instead of having to pass it via the EAX/RAX register as was the case with a breakpoint (and having to rely on the compiler to generate the proper assembly code).