Fix "Invalid parameter passed to C runtime function" debug output on WerTweak.dll injection
This output is being generated internally by LoadLibraryW() when its lpFileName parameter isn't 2-byte aligned. In our case, this parameter is a field in the assembly stub that gets written to WerFault.exe's memory to inject WerTweak.dll. Both the 32-bit and 64-bit assembly stub were affected. This was fixed by adding padding to both stubs so that their 'DLL name' field is now 2-byte aligned.